+++ src/presenters/pages/router.js
@@ -6,8 +6,6 @@ import categories from '../../curated/categories';
 import rootTeams from '../../curated/teams';
 
 import { useCurrentUser } from '../../state/current-user';
-import useLocalStorage from '../../state/local-storage';
-import { useAPI } from '../../state/api';
 
 import IndexPage from './index';
 import { FacebookLoginPage, GitHubLoginPage, GoogleLoginPage, EmailTokenLoginPage } from './login';
@@ -61,38 +59,8 @@ const PageChangeHandler = withRouter(({ location }) => {
   return null;
 });
 
-const SuperUserBanner = () => {
-  const { currentUser, persistentToken } = useCurrentUser();
-  const api = useAPI();
-  const [showSupportBanner, setShowSupportBanner] = useLocalStorage('showSupportBanner', false);
-  const isSupporter = currentUser && currentUser.projects && currentUser.projects.filter((p) => p.id === 'b9f7fbdd-ac07-45f9-84ea-d484533635ff').length > 0;
-
-  if (isSupporter && persistentToken) {
-    const superUser = currentUser.features && currentUser.features.find((feature) => feature.name === 'super_user');
-    const expirationDate = superUser && new Date(superUser.expiresAt).toUTCString();
-    const displayText = `SUPER USER MODE ${superUser ? `ENABLED UNTIL: ${expirationDate}` : 'DISABLED'} `;
-    const toggleSuperUser = async () => {
-      await api.post(`https://support-toggle.glitch.me/support/${superUser ? 'disable' : 'enable'}`);
-      window.location.reload();
-    };
-
-    if (superUser || showSupportBanner) {
-      return (
-        <div style={{ backgroundColor: `${superUser ? 'red' : 'aliceblue'}`, padding: '10px', textAlign: 'center', fontWeight: 'bold' }}>
-          {displayText}
-          <button onClick={toggleSuperUser}>Click to {superUser ? 'disable' : 'enable'}</button>
-          {!superUser && <button onClick={() => setShowSupportBanner(false)}>Hide</button>}
-        </div>
-      );
-    }
-  }
-
-  return null;
-};
-
 const Router = () => (
   <>
-    <SuperUserBanner />
     <PageChangeHandler />
     <Switch>
       <Route path="/" exact render={({ location }) => <IndexPage key={location.key} />} />
